pyproject.toml
[project]
name = "paperlens"
version = "0.1.0"
description = "Your personal research lens — AI-powered personalized paper tracking & summarization"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "apscheduler>=3.10.4",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "jinja2>=3.1.4",
  "pydantic-settings>=2.10.1",
  "python-dotenv>=1.1.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "sqlalchemy>=2.0.43",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
  "ruff>=0.12.12",
  "pre-commit>=4.0.0",
]
llm = [
  "openai>=1.102.0",
  "anthropic>=0.62.0",
]
pdf = [
  "pymupdf>=1.25.5",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
  "apps",
  "apps.api",
  "apps.desktop",
  "apps.worker",
  "packages",
  "packages.ai",
  "packages.domain",
  "packages.integrations",
  "packages.storage",
]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
  ".git",
  ".venv",
  "__pycache__",
  "node_modules",
  "data",
  "infra/migrations/versions",
]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # isort
  "UP",  # pyupgrade
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "SIM", # flake8-simplify
  "TCH", # flake8-type-checking
]
ignore = [
  "E501",  # line too long (handled by formatter)
  "B008",  # do not perform function calls in argument defaults
  "B905",  # zip without strict
  "SIM108", # use ternary operator
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "packages"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
